requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
filelock==3.13.1
fsspec==2023.12.2
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
pandas==2.2.0
pillow==10.2.0
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sympy==1.12
torch==2.2.0
torchvision==0.17.0
transforms==0.2.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0



### Environment setup for Maskformer2
```bash
pip install -q git+https://github.com/huggingface/transformers.git
pip install -q evaluate
pip install mmcv==1.3.14

```
